UntimedGame now inherits from Game superclass and uses its methods for everything except untimed game specific things like move counting. Some Checkstyle cleanup in SumFunFrame since I was commiting code anyway.
